Refactored util files. Tweaked gauges for greater arcs.
