+++ skins/weewx-wdc/skin.conf
@@ -165,22 +165,28 @@ SKIN_VERSION = 3.2.0
                 min = -20
                 max = 40
             [[[[dewpoint]]]]
+                arc = 1.1
             [[[[outHumidity]]]]
                 min = 0
                 max = 100
+                arc = 1.0
             [[[[barometer]]]]
                 offset = 3
+                arc = 0.9
             [[[[windSpeed]]]]
                 min = 0
             [[[[UV]]]]
                 min = 0
                 max = 12
+                arc = 1.3
             [[[[radiation]]]]
                 min = 0
+                arc = 1.5
             [[[[appTemp]]]]
             [[[[cloudbase]]]]
+                min = 0
             [[[[windDir]]]]
-                arc = 2.0
+                arc = 2
                 min = 0
                 max = 360
