+++ skins/weewx-wdc/src/js/diagrams/d3/line.tsx
@@ -24,7 +24,7 @@ import {
   getColors,
   getCurve,
   getObsPropsFromChartProps,
-} from "./components/util";
+} from "./../../util/util";
 import { addLegend } from "./components/legend";
 
 type LineDiagramBaseProps = DiagramBaseProps & {
